requirements.txt
docopt==0.6.2
stomp.py==6.0.0
websocket-client==1.7.0
